requirements.txt
asgiref
Django
djangorestframework
djangorestframework_simplejwt
drf-yasg
inflection
packaging
PyJWT
python-dotenv
pytz
PyYAML
sqlparse
tzdata
uritemplate
